custom-drawers: refactoring
