+++ Assets/OneLine/OneLine/Editor/Drawers/CustomDrawer/TypeForDrawing.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c9196356e41a84398866f3eedbd5939c
+timeCreated: 1519643676
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
